Add more delay after starting the opensm in Phase 1_1
When we start the opensm again after killing it, we need to wait a
little bit for the subnet manager to come up and configure the links,
otherwise the following tests may fail.
Let's wait a while until we continue in this case, so the following
testcase will not fail randomly.

Signed-off-by: Michael Moese mmoese@suse.de
Signed-off-by: Michael Moese <mmoese@suse.de>