Sort the output of ibstat -l
On one of my machines, ibstat returns the second port of a card before
the first one. Those machines are connected back to back, without any
switch, and actually only the first port is to be used.
When not sorting the ports, the ib_ipoib configuration is done in a way,
that will use the wrong interface and thus end up using routes that will
never reach anything.

Signed-off-by: Michael Moese <mmoese@suse.de>